Refactored sensor block readonly state data
